shramify1/requirements.txt
Flask==3.0.3
pymongo==4.8.0
werkzeug==3.0.3
python-dotenv==1.0.0
requests==2.31.0
twilio==8.10.0
razorpay==1.4.2
google-auth==2.34.0
Pillow==10.4.0
pytesseract==0.3.13
